+++ HISTORY.md
@@ -1,4 +1,7 @@
-v0.2.0 (dev)
+v0.2.1 (dev)
+============
+
+v0.2.0 
 ======
 + `indexcov`: better error message and handling of excluded chromosomes.
 + `indexcov`: if given crais and no fai, indexcov will try to read the cram header using `samtools view`.